% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capivara-package.R
\docType{package}
\name{capivara-package}
\alias{capivara}
\alias{capivara-package}
\title{capivara}
\description{
Spectral segmentation and analysis tools for IFU data cubes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rafaelsdesouza.com.br/capivara/}
  \item \url{https://github.com/RafaelSdeSouza/capivara}
  \item Report bugs at \url{https://github.com/RafaelSdeSouza/capivara/issues}
}

}
\keyword{internal}
