% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics_disc_model.R
\name{disc_velocity_model}
\alias{disc_velocity_model}
\title{Axisymmetric arctangent rotating-disc velocity model}
\usage{
disc_velocity_model(par, R, theta, inc_rad)
}
\arguments{
\item{par}{Named vector/list with `vsys`, `vmax`, and `Rt`.}

\item{R}{Deprojected radius in spaxels.}

\item{theta}{Deprojected azimuth.}

\item{inc_rad}{Inclination in radians.}
}
\value{
Model line-of-sight velocity.
}
\description{
Axisymmetric arctangent rotating-disc velocity model
}
