% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics_plot_capivara_kinematics.R
\name{plot_capivara_kinematics}
\alias{plot_capivara_kinematics}
\title{Plot Capivara segmentation-aware kinematic modelling products}
\usage{
plot_capivara_kinematics(result, png_file = NULL, pdf_file = NULL)
}
\arguments{
\item{result}{Result from `run_one_galaxy()`.}

\item{png_file, pdf_file}{Output paths.}
}
\value{
The patchwork plot object.
}
\description{
Plot Capivara segmentation-aware kinematic modelling products
}
