% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spec_mean.R
\name{Spec_mean}
\alias{Spec_mean}
\title{Compute Median Spectra and Process Clustered Data from a Data Cube}
\usage{
Spec_mean(cluster_result)
}
\arguments{
\item{input_cube}{A FITS file or object representing the input data cube.
Typically, this is the original IFU data cube used in clustering.}

\item{cluster_output}{A list returned by the \code{\link{cube_cluster}} function,
containing the cluster map (\code{cluster_map}), metadata (\code{header}, \code{axDat}),
and SNR estimates (\code{cluster_snr}).}

\item{output_dir}{A character string specifying the directory where the output files
will be saved. Defaults to \code{"output_results"}.}
}
\value{
A list containing:
\item{MeanSpec}{A data frame with the median spectrum for each cluster, where rows correspond to clusters and columns correspond to wavelengths.}
\item{median_cube_file}{The file path to the FITS file containing the new cube with median spectra assigned to each pixel.}
\item{post_process_table}{The file path to the CSV file containing the table of original spectra and their associated clusters.}
}
\description{
This function processes a data cube and the corresponding clustering output to compute
the median spectrum for each cluster, generate a new cube with median spectra assigned
to each pixel, and create a table linking each pixel's original spectrum to its cluster.
}
\details{
Steps performed by the function:
\enumerate{
  \item Reads the input data cube and clustering output.
  \item Reshapes the data cube and cluster map into 2D formats.
  \item Computes the median spectrum for each cluster by grouping pixels based on their cluster assignment.
  \item Creates a new cube where each pixel's spectrum is replaced by the median spectrum of its cluster.
  \item Generates a table linking each pixel's original spectrum to its assigned cluster.
  \item Saves the median spectrum cube and the table to the specified directory.
}

This function is designed to facilitate post-processing and analysis of clustered
spectral data, commonly used in IFU data analysis to study regions with similar
spectral properties.
}
\examples{
# Example usage with a FITS file and cluster output
cube_cluster_result <- cube_cluster(
  input = FITSio::readFITS("manga-11749-12701-LOGCUBE.fits"),
  Ncomp = 5
)

Spec_mean_result <- Spec_mean(
  input_cube = FITSio::readFITS("manga-11749-12701-LOGCUBE.fits"),
  cluster_output = cube_cluster_result
)

# Access outputs
MeanSpec <- Spec_mean_result$MeanSpec
median_cube_file <- Spec_mean_result$median_cube_file
post_process_table <- Spec_mean_result$post_process_table

}
\seealso{
\code{\link{cube_cluster}}, \code{\link{FITSio}}, \code{\link{reshape2}}, \code{\link{dplyr}}
}
